feat: ho provato a fare qualcosa ma si fa fatica, ho fatto ordine in app.jsx
