+++ Client/components/searchbar/Searchbar.jsx
@@ -4,21 +4,9 @@ import { Link } from "react-router-dom";
 import GlobalContext from "../../context/GlobalContext";
 
 function SearchBar() {
-    const { setSlugDoctor, allSpec, selectedSpec, setSelectedSpec, setNameSpecSelected } = useContext(GlobalContext);
-    const backurl = import.meta.env.VITE_BACKEND_URL;
 
-    const [filters, setFilters] = useState({
-        firstname: "",
-        lastname: "",
-        specialization: "",
-    });
-
-    const [doctors, setDoctors] = useState([]);
-    const [loading, setLoading] = useState(false);
-    const [error, setError] = useState("");
-    const [page, setPage] = useState(1);
-    const [limit] = useState(10);
-    const [searching, setSearching] = useState(false);
+    // context
+    const { allSpec, filters, loading, error, page, searching, handleInputChange, handleSubmit, searchDoctors, handleSelect } = useContext(GlobalContext);   
 
     // Quando i filters o la pagina cambiano, resettiamo i medici
     useEffect(() => {
@@ -27,57 +15,12 @@ function SearchBar() {
         }
     }, [filters, page, searching]);
 
-    const handleInputChange = (event) => {
-        const { name, value } = event.target;
-        setFilters((prevFilters) => ({
-            ...prevFilters,
-            [name]: value,
-        }));
-    };
-
-    const handleSubmit = (event) => {
-        event.preventDefault();
-        setDoctors([]);  // Reset lista medici
-        setPage(1);  // Reset della pagina a 1
-        setSearching(true);  // Abilita la ricerca
-    };
-
-    const searchDoctors = async () => {
-        setLoading(true);
-        setError("");
-        try {
-            const response = await axios.get(`${backurl}doctors`, {
-                params: { ...filters, page, limit },
-            });
-
-            setDoctors(response.data.data);
-        } catch (error) {
-            setError("Errore nella ricerca dei dottori.");
-        } finally {
-            setLoading(false);
-        }
-    };
-
-    const handlePageChange = (newPage) => {
-        if (newPage >= 1) {
-            setPage(newPage);
-        }
-    };
-
-    // Disabilitare il pulsante "Successivo" se non ci sono più medici
-    const disableNextButton = doctors.length < limit;
-
-    const handleSelectSpecialization = (event) => {
-        setFilters((prevFilters) => ({
-            ...prevFilters,
-            specialization: event.target.value,
-        }));
-    };
-
     return (
         <div className="container mt-4">
             <h1>Ricerca Dottori</h1>
-            <form onSubmit={handleSubmit} className="d-flex gap-3 mb-4">
+            <form onSubmit={() => handleSubmit(event)} className="d-flex gap-3 mb-4">
+
+
                 {/* Nome Input */}
                 <div className="mb-3 w-25">
                     <label htmlFor="firstname" className="form-label">
@@ -88,7 +31,7 @@ function SearchBar() {
                         id="firstname"
                         name="firstname"
                         value={filters.firstname}
-                        onChange={handleInputChange}
+                        onChange={() => handleInputChange(event)}
                         placeholder="Nome"
                         className="form-control"
                     />
@@ -104,7 +47,7 @@ function SearchBar() {
                         id="lastname"
                         name="lastname"
                         value={filters.lastname}
-                        onChange={handleInputChange}
+                        onChange={() => handleInputChange(event)}
                         placeholder="Cognome"
                         className="form-control"
                     />
@@ -118,7 +61,7 @@ function SearchBar() {
                     <select
                         name="specialization"
                         value={filters.specialization}
-                        onChange={handleSelectSpecialization}
+                        onChange={() => handleSelect(event)}
                         className="form-control"
                     >
                         <option value="">-- Seleziona una specializzazione --</option>
@@ -143,7 +86,7 @@ function SearchBar() {
 
             {error && <div className="alert alert-danger">{error}</div>}
 
-            <div>
+            {/* <div>
                 {doctors.length === 0 && !loading && searching && <p>Nessun dottore trovato.</p>}
                 {searching && (
                     <ul className="list-group">
@@ -173,7 +116,7 @@ function SearchBar() {
                 >
                     Successivo
                 </button>
-            </div>
+            </div> */}
         </div>
     );
 }
